Rename utils module and update imports
